+++ js/main.js
@@ -1,5 +1,3 @@
-import './createPhoto.js';
-// import { photo } from './createPhoto.js';
 import './miniatures.js';
 import './renderPhoto.js';
 import './uploadImage.js';
